Move essential parts of user registration out of the controller.
